Fix koin bugs, request going through

+++ data/src/main/java/com/zistus/data/datasources/db/room/DatabaseSource.kt
@@ -5,5 +5,5 @@ import io.reactivex.Single
 
 interface DatabaseSource {
 
-    fun insertItem(item: Entity.TestObject): Single<Int>
+    fun insertItem(item: Entity.TestObject): Single<Long>
 }
